Flattened classes structure. Introduced abstract Linear creative.
